requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
python-dateutil